% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.subsamples.R
\name{ccc}
\alias{ccc}
\title{calculate concordance correlation coefficient}
\usage{
ccc(x, y)
}
\arguments{
\item{x}{first vector}

\item{y}{second vector}
}
\value{
concordance correlation coefficient
}
\description{
calculate concordance correlation coefficient
}
\details{
uses only complete observations
}

