% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ccc}
\alias{ccc}
\title{calculate concordance correlation coefficient}
\usage{
ccc(x, y)
}
\arguments{
\item{x}{first vector}

\item{y}{second vector}
}
\description{
calculate concordance correlation coefficient
}
\details{
uses only complete observations
}

