% Generated by roxygen2 (4.0.2): do not edit by hand
\name{DEXSeq}
\alias{DEXSeq}
\title{handler for performing DEXSeq on subsampled RNA-Seq data}
\usage{
DEXSeq(count.matrix, design, geneIDs, exonIDs)
}
\arguments{
\item{count.matrix}{Matrix of unnormalized counts. The row names should
represent the exon names}

\item{design}{Study design data frame for DEXSeq}

\item{geneIDs}{vector of gene IDs- one for each row of the count matrix
(repeated for each exon in a gene)}

\item{exonIDs}{vector of exon IDs- one for each row of the count matrix}
}
\value{
A data frame with one row for each gene. This contains all columns
from the \link{DEXSeq::DEXSeqResults} function in DEXSeq, except that the
\code{log2fold_untreated_treated} column is been changed to \code{coefficient}
and an \code{ID} column is added, with the format "geneID:exonID"
}
\description{
This encapsulates the DEXSeq dispersion estimation and \code{testForDEU}
functions. It takes a count matrix, potentially subsampled, and a data table
representing the study design, and tests for differential exon usage.
}
\examples{
data( "pasillaExons", package="pasilla" )
exon.counts = pasillaExons@assayData$counts

design = pasillaExons@phenoData@data[c("condition", "type")]
geneIDs = pasillaExons@featureData@data$geneID
exonIDs = pasillaExons@featureData@data$exonID

DEXSeq(exon.counts, design, geneIDs, exonIDs)
}

